==> 00_core/01_clocks/05_visualize/02_GoTermViz.R <==
# GO Analysis for Clock Specific and Clock Shared Genes

library(tidyverse)
library(viridis)

setwd("~/Dropbox/svz_singlecell_aging_clocks/00_core/01_clocks/05_visualize")

# Table GO_Biological_Process_2021_table.csv generated by manually entering cell-type
# specific clock genes (SpecificAndSharedGenes.xlsx) into enrichr website and collating the results.

go <- read.csv("data/GO_Biological_Process_2021_table.csv")
colnames(go)[1] <- "Term"
go <- tibble(go)
go$Celltype <- factor(go$Celltype, levels = c( "Oligodendrocyte", "Microglia", "Endothelial", "Astrocyte_qNSC", "aNSC_NPC", "Neuroblast", "Shared"))

# Top N highest values by group
go2 <- go %>%                                      
  arrange(desc(-P.value)) %>% 
  group_by(Celltype) %>%
  slice(1:4)

terms_interest <- go2$Term
go3 <- go[go$Term %in% terms_interest, ]
go3$Term <- factor(go3$Term, levels = as.character(rev(go2$Term)))


ggplot(go3, aes(y = Term, x = Celltype, size = -log(P.value), color = -log(P.value))) +
    geom_point() +
    scale_color_viridis(option = "viridis", begin = 0, end = .99, direction = 1) +
    #scale_color_distiller(palette="RdYlGn") +
    geom_point(shape = 1, color = "black") +
    theme_bw() +
    theme(axis.title.x = element_blank(), axis.title.y = element_blank()) +
    theme(axis.text.x = element_text(angle = 30, hjust=1))
ggsave("plots/GOBP_enricher.pdf", useDingbats=F, width =8.75, height = 4.5)


#=====================================================================================
# Repeat with genes from bootstarp bio clock

go <- read.csv("data/GO_Biological_Process_2021_table_bio.csv")
colnames(go)[1] <- "Term"
go <- tibble(go)
go$Celltype <- factor(go$Celltype, levels = c("Oligodendrocyte", "Microglia", "Endothelial", "Astrocyte_qNSC", "aNSC_NPC", "Neuroblast", "Shared"))

# Top N highest values by group
go2 <- go %>%                                      
  arrange(desc(-P.value)) %>% 
  group_by(Celltype) %>%
  slice(1:4)

terms_interest <- go2$Term
go3 <- go[go$Term %in% terms_interest, ]
go3$Term <- factor(go3$Term, levels = unique(as.character(rev(go2$Term))))

ggplot(go3, aes(y = Term, x = Celltype, size = -log(P.value), color = -log(P.value))) +
    geom_point() +
    scale_color_viridis(option = "viridis", begin = 0, end = .99, direction = 1) +
    geom_point(shape = 1, color = "black") +
    theme_bw() +
    theme(axis.title.x = element_blank(), axis.title.y = element_blank()) +
    theme(axis.text.x = element_text(angle = 30, hjust=1))
ggsave("plots/enricher_bio_Sep8.pdf", useDingbats=F, width =8.75, height = 4.5)
